Refactor: FeedResponse의 필드 변경
