+++ src/main/java/com/stoury/dto/feed/FeedResponse.java
@@ -3,14 +3,13 @@ package com.stoury.dto.feed;
 import com.stoury.domain.Feed;
 import com.stoury.domain.GraphicContent;
 import com.stoury.domain.Tag;
-import com.stoury.dto.LocationResponse;
-import com.stoury.dto.member.MemberResponse;
+import com.stoury.dto.WriterResponse;
 
 import java.time.LocalDateTime;
 import java.util.List;
 
 public record FeedResponse(Long feedId,
-                           MemberResponse memberResponse,
+                           WriterResponse writer,
                            List<String> graphicContentsPaths,
                            String textContent,
                            Double latitude,
@@ -22,7 +21,7 @@ public record FeedResponse(Long feedId,
     public static FeedResponse from(Feed feed, long feedLikes) {
         return new FeedResponse(
                 feed.getId(),
-                MemberResponse.from(feed.getMember()),
+                WriterResponse.from(feed.getMember()),
                 feed.getGraphicContents().stream().map(GraphicContent::getPath).toList(),
                 feed.getTextContent(),
                 feed.getLatitude(),
