+++ src/main/java/com/stoury/dto/feed/LocationResponse.java
@@ -1,4 +1,4 @@
-package com.stoury.dto;
+package com.stoury.dto.feed;
 
 public record LocationResponse(String city, String country) {
 }
